requirements.txt
grpcio~=1.47.0
protobuf~=3.19.1

numpy~=1.23.1